setting project configuration
