+++ app/build.gradle
@@ -44,7 +44,14 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-
     //materialDesign
     implementation 'com.google.android.material:material:1.6.0-alpha02'
+    // ViewModel
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.4.1'
+    // LiveData
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.4.1'
+    // Fragment
+    implementation 'androidx.fragment:fragment-ktx:1.4.1'
+    // Activity
+    implementation 'androidx.activity:activity-ktx:1.4.0'
 }
