+++ app/src/main/java/com/example/bankodemia/UI/View/DataFragment.kt
@@ -1,10 +1,11 @@
-package com.example.bankodemia
+package com.example.bankodemia.UI.View
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.bankodemia.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -13,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [TelephoneFragment.newInstance] factory method to
+ * Use the [DataFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class TelephoneFragment : Fragment() {
+class DataFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +35,7 @@ class TelephoneFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_telephone, container, false)
+        return inflater.inflate(R.layout.fragment_data, container, false)
     }
 
     companion object {
@@ -44,12 +45,12 @@ class TelephoneFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment TelephoneFragment.
+         * @return A new instance of fragment DataFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            TelephoneFragment().apply {
+            DataFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
