+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/black"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".UI.View.MainActivity">
 
     <ScrollView
         android:layout_width="0dp"
