+++ app/src/main/res/layout/fragment_confirmation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ConfirmationFragment">
+    tools:context=".UI.View.ConfirmationFragment">
 
     <ImageView
         android:id="@+id/confirmationIvLogo"
