+++ app/src/main/res/layout/fragment_contact_added.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactAddedFragment">
+    tools:context=".UI.View.ContactAddedFragment">
 
     <ImageView
         android:id="@+id/imageView4"
