+++ app/src/main/res/layout/fragment_create_account.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateAccountFragment">
+    tools:context=".UI.View.CreateAccountFragment">
 
     <ScrollView
         android:layout_width="match_parent"
