+++ app/src/main/res/layout/fragment_data.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DataFragment">
+    tools:context=".UI.View.DataFragment">
 
     <ScrollView
         android:layout_width="match_parent"
