+++ app/src/main/res/layout/fragment_identity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IdentityFragment">
+    tools:context=".UI.View.IdentityFragment">
 
     <ImageView
         android:id="@+id/identityIvLogo"
