+++ app/src/main/res/layout/fragment_identity_menu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IdentityMenuFragment">
+    tools:context=".UI.View.IdentityMenuFragment">
 
     <ImageView
         android:id="@+id/identityMenuIvLogo"
