+++ app/src/main/res/layout/fragment_ine.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".IneFragment">
+    tools:context=".UI.View.IneFragment">
 
     <ImageView
         android:id="@+id/ineIvLogo"
