+++ app/src/main/res/layout/fragment_migratory.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MigratoryFragment">
+    tools:context=".UI.View.MigratoryFragment">
 
     <ImageView
         android:id="@+id/migratoryIvLogo"
