+++ app/src/main/res/layout/fragment_sending_data.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SendingDataFragment">
+    tools:context=".UI.View.SendingDataFragment">
 
     <ImageView
         android:id="@+id/sendingDataIvLogo"
