+++ app/src/main/res/layout/fragment_telephone.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TelephoneFragment">
+    tools:context=".UI.View.TelephoneFragment">
 
     <ScrollView
         android:layout_width="match_parent"
