+++ app/src/main/res/layout/fragment_waiting_transaction.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WaitingTransactionFragment">
+    tools:context=".UI.View.WaitingTransactionFragment">
 
 
     <ImageView
